NumberRecognizer.add_format supports :country_code parameter, added prefix
When using the default NumerRecognizer, country and prefix are the same.
When using custom formats, you can add a :country_code which overrides the country value returned by number recognizer.
The use case I have for this is when we have fake msisdns generated in a 89x prefix range (officially not assigned to any country), we can still say that it belongs to a Dutch person with country code 31.